+++ crates/sncast/src/helpers/fee.rs
@@ -5,11 +5,12 @@ use conversions::TryIntoConv;
 use starknet::core::types::BlockId;
 use starknet::providers::Provider;
 use starknet_types_core::felt::{Felt, NonZeroFelt};
+use std::str::FromStr;
 
 #[derive(Args, Debug, Clone)]
 pub struct FeeArgs {
     /// Token that transaction fee will be paid in
-    #[clap(long)]
+    #[clap(long, value_parser = parse_fee_token)]
     pub fee_token: Option<FeeToken>,
 
     /// Max fee for the transaction. If not provided, will be automatically estimated.
@@ -50,9 +51,9 @@ impl From<ScriptFeeSettings> for FeeArgs {
 
 impl FeeArgs {
     #[must_use]
-    pub fn fee_token(self, fee_token: Option<FeeToken>) -> Self {
+    pub fn fee_token(self, fee_token: FeeToken) -> Self {
         Self {
-            fee_token: fee_token.or(self.fee_token),
+            fee_token: Some(fee_token),
             ..self
         }
     }
@@ -138,9 +139,10 @@ impl FeeArgs {
     }
 }
 
-#[derive(ValueEnum, Debug, Clone, PartialEq)]
+#[derive(ValueEnum, Default, Debug, Clone, PartialEq)]
 pub enum FeeToken {
     Eth,
+    #[default]
     Strk,
 }
 
@@ -187,7 +189,7 @@ impl From<ScriptFeeSettings> for FeeSettings {
 
 pub trait PayableTransaction {
     fn error_message(&self, token: &str, version: &str) -> String;
-    fn validate(&self) -> Result<()>;
+    fn validate_and_get_token(&self) -> Result<FeeToken>;
     fn token_from_version(&self) -> Option<FeeToken>;
 }
 
@@ -199,21 +201,27 @@ macro_rules! impl_payable_transaction {
                 $err_func(token, version)
             }
 
-            fn validate(&self) -> Result<()> {
+            fn validate_and_get_token(&self) -> Result<FeeToken> {
                 match (
                     &self.token_from_version(),
                     &self.fee_args.fee_token,
                 ) {
                     (Some(token_from_version), Some(token)) if token_from_version != token => {
                         Err(anyhow!(self.error_message(
-                            &format!("{token:?}").to_lowercase(),
+                            &format!("{:?}", token).to_lowercase(),
                             &format!("{:?}", self.version.clone().unwrap()).to_lowercase()
                         )))
-                    }
+                    },
+                    (None, Some(token)) => {
+                        Ok(token.clone())
+                    },
+                    (Some(token_from_version), None) => {
+                        Ok(token_from_version.clone())
+                    },
                     (None, None) => {
-                        Err(anyhow!("Either --fee-token or --version must be provided"))
-                    }
-                    _ => Ok(()),
+                        Ok(FeeToken::default())
+                    },
+                    _ =>  Ok(self.fee_args.fee_token.clone().unwrap_or_else(|| self.token_from_version().unwrap_or_else(|| unreachable!())))
                 }
             }
 
@@ -224,4 +232,30 @@ macro_rules! impl_payable_transaction {
             }
         }
     };
+}
+
+impl FromStr for FeeToken {
+    type Err = String;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        match s.to_lowercase().as_str() {
+            "eth" => Ok(FeeToken::Eth),
+            "strk" => Ok(FeeToken::Strk),
+            _ => Err(String::from("Possible values: eth, strk")),
+        }
+    }
+}
+
+fn parse_fee_token(s: &str) -> Result<FeeToken, String> {
+    let deprecation_message = "\x1b[33mwarning:\x1b[0m Specifying '--fee-token' flag is deprecated and will be removed in the future. Use '--version' instead";
+
+    match s.to_lowercase().as_str() {
+        "eth" => {
+            println!("{deprecation_message}");
+            println!("\x1b[33mwarning:\x1b[0m Eth transactions will stop being supported in the future due to 'SNIP-16'");
+        }
+        _ => println!("{deprecation_message}"),
+    }
+    let parsed_token: FeeToken = s.parse()?;
+    Ok(parsed_token)
 }