+++ crates/sncast/tests/e2e/account/deploy.rs
@@ -349,7 +349,7 @@ async fn test_invalid_version_and_token_combination(fee_token: &str, version: &s
 }
 
 #[tokio::test]
-async fn test_no_version_and_token() {
+async fn test_default_fee_token() {
     let tempdir = create_account(false, &OZ_CLASS_HASH.into_hex_string(), "oz").await;
     let accounts_file = "accounts.json";
 
@@ -367,11 +367,14 @@ async fn test_no_version_and_token() {
 
     let snapbox = runner(&args).current_dir(tempdir.path());
 
-    let output = snapbox.assert().failure();
-    assert_stderr_contains(
-        output,
-        "Error: Either --fee-token or --version must be provided",
-    );
+    snapbox.assert().success().stdout_matches(indoc! {r"
+        Transaction hash: [..]
+        command: account deploy
+        transaction_hash: [..]
+
+        To see invocation details, visit:
+        transaction: [..]
+    "});
 }
 
 #[tokio::test]
@@ -390,8 +393,6 @@ pub async fn test_valid_class_hash() {
         "my_account",
         "--max-fee",
         "10000000000000000",
-        "--fee-token",
-        "eth",
     ];
 
     let snapbox = runner(&args).current_dir(tempdir.path());
@@ -421,8 +422,6 @@ pub async fn test_valid_no_max_fee() {
         URL,
         "--name",
         "my_account",
-        "--fee-token",
-        "eth",
     ];
 
     let snapbox = runner(&args).current_dir(tempdir.path());
@@ -515,8 +514,6 @@ pub async fn test_happy_case_keystore(account_type: &str) {
         URL,
         "--max-fee",
         "99999999999999999",
-        "--fee-token",
-        "eth",
     ];
 
     let snapbox = runner(&args).current_dir(tempdir.path());
@@ -791,8 +788,6 @@ pub async fn test_deploy_keystore_other_args() {
         "some-name",
         "--max-fee",
         "99999999999999999",
-        "--fee-token",
-        "eth",
     ];
 
     let snapbox = runner(&args).current_dir(tempdir.path());